+++ ai-service/src/main/java/com/firstlogistics/aiservice/application/dto/result/AILogResult.java
@@ -1,0 +1,21 @@
+package com.firstlogistics.aiservice.application.dto.result;
+
+import com.firstlogistics.aiservice.domain.entity.AILog;
+import com.firstlogistics.aiservice.domain.enums.AILogStatus;
+import com.firstlogistics.aiservice.domain.enums.MessengerType;
+
+import java.util.UUID;
+
+public record AILogResult(
+        UUID id,
+        AILogStatus status,
+        MessengerType messengerType
+) {
+    public static AILogResult from(AILog aiLog) {
+        return new AILogResult(
+                aiLog.getId().id(),
+                aiLog.getStatus(),
+                aiLog.getMessengerType()
+        );
+    }
+}
