+++ ai-service/src/main/java/com/firstlogistics/aiservice/application/external/AIConfig.java
@@ -1,0 +1,14 @@
+package com.firstlogistics.aiservice.application.external;
+
+import org.springframework.ai.chat.client.ChatClient;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AIConfig {
+
+    @Bean
+    public ChatClient chatClient(ChatClient.Builder builder) {
+        return builder.build();
+    }
+}
