+++ ai-service/src/main/java/com/firstlogistics/aiservice/application/external/AIPromptGenerator.java
@@ -1,0 +1,10 @@
+package com.firstlogistics.aiservice.application.external;
+
+import com.firstlogistics.aiservice.application.dto.command.CreateAILogCommand;
+
+public interface AIPromptGenerator {
+
+    String generateDeliveryGuide(CreateAILogCommand command);
+
+    String getSystemPrompt();
+}
