+++ ai-service/src/main/java/com/firstlogistics/aiservice/domain/enums/MessengerType.java
@@ -1,0 +1,5 @@
+package com.firstlogistics.aiservice.domain.enums;
+
+public enum MessengerType {
+    SLACK
+}
