+++ ai-service/src/main/java/com/firstlogistics/aiservice/domain/exception/AILogException.java
@@ -1,0 +1,9 @@
+package com.firstlogistics.aiservice.domain.exception;
+
+import common.exception.BaseException;
+
+public class AILogException extends BaseException {
+    public AILogException(AILogErrorCode message) {
+        super(message);
+    }
+}
