+++ ai-service/src/main/java/com/firstlogistics/aiservice/domain/repository/AILogRepository.java
@@ -1,0 +1,8 @@
+package com.firstlogistics.aiservice.domain.repository;
+
+import com.firstlogistics.aiservice.domain.entity.AILog;
+
+public interface AILogRepository {
+
+    AILog save(AILog entity);
+}
