+++ ai-service/src/main/java/com/firstlogistics/aiservice/infrastructure/persistence/jpa/AILogJpaRepository.java
@@ -1,0 +1,8 @@
+package com.firstlogistics.aiservice.infrastructure.persistence.jpa;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface AILogJpaRepository extends JpaRepository<AILogJpaEntity, UUID> {
+}
