--- notification-service/src/main/java/com/firstlogistics/notificationservice/ailog/application/dto/command/CreateAILogCommand.java
@@ -1,41 +1,0 @@
-package com.firstlogistics.notificationservice.ailog.application.dto.command;
-
-import com.firstlogistics.notificationservice.ailog.domain.enums.AILogStatus;
-import com.firstlogistics.notificationservice.ailog.domain.enums.MessengerType;
-
-import java.util.UUID;
-
-public record CreateAILogCommand(
-        UUID orderId,
-        String customerName,
-        String customerEmail,
-        String orderTime,
-        String productInfo,
-        String requirement,
-        String departure,
-        String hubs,
-        String destination,
-        String deliveryStaffName,
-        String deliveryStaffEmail,
-        String hubManagerSlackId
-) {
-
-
-    // Generator에서 사용하기 편하게 텍스트로 변환하는 헬퍼 메서드
-    public String toRawText() {
-        return String.format("""
-                주문 번호 : %s
-                주문자 정보 : %s / %s
-                주문 시간 : %s
-                상품 정보 : %s
-                요청 사항 : %s
-                발송지 : %s
-                경유지 : %s
-                도착지 : %s
-                배송담당자 : %s / %s
-                """,
-                orderId, customerName, customerEmail, orderTime,
-                productInfo, requirement, departure, hubs,
-                destination, deliveryStaffName, deliveryStaffEmail);
-    }
-}
