--- notification-service/src/main/java/com/firstlogistics/notificationservice/ailog/application/external/AIConfig.java
@@ -1,14 +1,0 @@
-package com.firstlogistics.notificationservice.ailog.application.external;
-
-import org.springframework.ai.chat.client.ChatClient;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class AIConfig {
-
-    @Bean
-    public ChatClient chatClient(ChatClient.Builder builder) {
-        return builder.build();
-    }
-}
