--- notification-service/src/main/java/com/firstlogistics/notificationservice/ailog/application/external/AIPromptGenerator.java
@@ -1,10 +1,0 @@
-package com.firstlogistics.notificationservice.ailog.application.external;
-
-import com.firstlogistics.notificationservice.ailog.application.dto.command.CreateAILogCommand;
-
-public interface AIPromptGenerator {
-
-    String generateDeliveryGuide(CreateAILogCommand command);
-
-    String getSystemPrompt();
-}
