--- notification-service/src/main/java/com/firstlogistics/notificationservice/ailog/domain/enums/MessengerType.java
@@ -1,5 +1,0 @@
-package com.firstlogistics.notificationservice.ailog.domain.enums;
-
-public enum MessengerType {
-    SLACK
-}
