--- notification-service/src/main/java/com/firstlogistics/notificationservice/ailog/domain/exception/AILogException.java
@@ -1,9 +1,0 @@
-package com.firstlogistics.notificationservice.ailog.domain.exception;
-
-import common.exception.BaseException;
-
-public class AILogException extends BaseException {
-    public AILogException(AILogErrorCode message) {
-        super(message);
-    }
-}
