--- notification-service/src/main/java/com/firstlogistics/notificationservice/ailog/domain/repository/AILogRepository.java
@@ -1,8 +1,0 @@
-package com.firstlogistics.notificationservice.ailog.domain.repository;
-
-import com.firstlogistics.notificationservice.ailog.domain.entity.AILog;
-
-public interface AILogRepository {
-
-    AILog save(AILog entity);
-}
