--- notification-service/src/main/java/com/firstlogistics/notificationservice/ailog/domain/vo/AILogId.java
@@ -1,13 +1,0 @@
-package com.firstlogistics.notificationservice.ailog.domain.vo;
-
-import java.util.UUID;
-
-public record AILogId(UUID id) {
-    public static AILogId of() {
-        return AILogId.of(UUID.randomUUID());
-    }
-
-    public static AILogId of(UUID id){
-        return new AILogId(id);
-    }
-}
