--- notification-service/src/main/java/com/firstlogistics/notificationservice/ailog/domain/vo/MessengerMessageId.java
@@ -1,13 +1,0 @@
-package com.firstlogistics.notificationservice.ailog.domain.vo;
-
-import java.util.Objects;
-import java.util.UUID;
-
-public record MessengerMessageId(UUID id) {
-    public MessengerMessageId {
-        Objects.requireNonNull(id, "MessengerMessageId must not be null");
-    }
-    public static MessengerMessageId of(UUID id) {
-        return new MessengerMessageId(id);
-    }
-}
