--- notification-service/src/main/java/com/firstlogistics/notificationservice/ailog/infrastructure/persistence/jpa/AILogJpaRepository.java
@@ -1,8 +1,0 @@
-package com.firstlogistics.notificationservice.ailog.infrastructure.persistence.jpa;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface AILogJpaRepository extends JpaRepository<AILogJpaEntity, UUID> {
-}
