--- notification-service/src/main/java/com/firstlogistics/notificationservice/ailog/presentation/AILogApiController.java
@@ -1,30 +1,0 @@
-package com.firstlogistics.notificationservice.ailog.presentation;
-
-import com.firstlogistics.notificationservice.ailog.application.service.AILogService;
-import com.firstlogistics.notificationservice.ailog.presentation.dto.request.AILogCreateRequest;
-import com.firstlogistics.notificationservice.ailog.presentation.dto.response.AILogResponse;
-import common.response.ApiResponse;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/ai-logs")
-public class AILogApiController {
-    private final AILogService aiLogService;
-
-    @PostMapping
-    public ResponseEntity<ApiResponse<AILogResponse>> createAILog(
-            @RequestBody @Valid AILogCreateRequest request
-    ) {
-        AILogResponse result = AILogResponse.from(aiLogService.createAILog(request.toCommand()));
-        return ResponseEntity.status(AILogSuccessCode.AILOG_CREATED.getStatus())
-                .body(ApiResponse.success(AILogSuccessCode.AILOG_CREATED, result));
-    }
-
-}
